Subject: [PATCH 1/2] EventDispatcher interface
The goal is to be able to support:

* PluggableUSB hosts
* Other USB implementations (such as Teensy 3.x)
* BLE on nRF52
* BLE via Adafruit SPIFRIEND boards
* Be able to direct keypresses to some combination of the above,
  depending on the policy in the firmware.  For example, the
  most desirable default would be to use USB when available,
  and fall back to BLE if available.

To that end, this diff adds:

* `PtrArray` - a reusable implementation of the same approach
  used for recording hooks, this is a bit like std::array or
  std::vector, but with the append/replace semantics from
  the existing hook registration.  A later commit could
  switch the hooks over to PtrArray.

* `EventDispatcher` - a virtual base class with methods
  relating to pressing different kinds of keys.  There
  is a global registry of `EventDispatcher`'s maintained
  using a `PtrArray`.  Each `EventDispatcher` registers
  itself during construction.

`PtrArray::apply` allows walking over the list of dispatchers and
applying a lambda.  If you're worried about the overhead of this, the
compiler should optimize this to the equivalent for loop and direct
function call.

The `EventDispatcher` has the concept of a `connectionMask` which
is essentially a bitfield that indicates the combination of USB
and BLE connections that are available.   Each dispatcher will
report its nature via the `queryConnectionTypes` method.

The loop driver then has the option to apply the policy mentioned
above.  I haven't provided the code for that portion, because
I haven't actually implemented that yet, and it can be done
in a separate smaller commit.  The key concept here is that
we have the option to transform the *available* `connectionMask`
into a *desired* `connectionMask`.

Each of the key handling related methods is passed the *desired*
`connectionMask` along with the information about the key press.
The implementation will NOP if the desired `connectionMask` mask
doesn't match its nature.

On Registration:

The intent is that the hardware class will simply embed an
instance of the appropriate `EventDispatcher`(s) as member fields.
This will implicitly register them.

The `HIDEventDispatcher.h` header contains an implementation
with the same semantics as we had prior to this diff, so most
USB based boards will just need to add:

```
   HIDEventDispatcher dispatcher_;
```

as a field and be done.

In an attempt to minimize the number of changes in the other repos,
I've added a little bit of preprocessor stuff to allow setting
up a default HIDEventDispatcher instance.  The protocol for this
is that, until the hardware classes are updated, if the hardware
header doesn't define `HARDWARE_EVENT_DISPATCHER`, the
`DefaultEventDispatcher.cpp` class will instantiate one.  The
`Kaleidoscope::setup` method will explicitly reference that class
to ensure that the linker pulls it in.

Fan out to other repos:

The Macro and DualKey repos each need a small change to use the
new `send_keyboard_report` functon instead of trying to call
`Keyboard.sendReport` directly.

Mouse Keys:

We can either expand `EventDispatcher` to incorporate mouse
related methods, or we can add a similar `MouseDispatcher` class
that works in basically the same way.  Thoughts?